// src/mangadex/api/structures/User/UserRole.ts
/**
 * [User Role Enum](https://api.mangadex.org/docs/static-data/#user-roles-enum)
 * 
 */
export enum UserRole {
    Admin = "ROLE_ADMIN",
    Banned = "ROLE_BANNED",
    Contributor = "ROLE_CONTRIBUTOR",
    Designer = "ROLE_DESIGNER",
    Developer = "ROLE_DEVELOPER",
    Forum_Moderator = "ROLE_FORUM_MODERATOR",
    Global_Moderator = "ROLE_GLOBAL_MODERATOR",
    Group_Leader = "ROLE_GROUP_LEADER",
    Group_Member = "ROLE_GROUP_MEMBER",
    Guest = "ROLE_GUEST",
    Member = "ROLE_MEMBER",
    Md_At_Home = "ROLE_MD_AT_HOME",
    Power_Uploader = "ROLE_POWER_UPLOADER",
    Public_Relations = "ROLE_PUBLIC_RELATIONS",
    Staff = "ROLE_STAFF",
    Unverified = "ROLE_UNVERIFIED",
    User = "ROLE_USER",
    VIP = "ROLE_VIP",
    Unknown = "ROLE_UNKNOWN"
}